requirements.txt

certifi==2018.4.16
chardet==3.0.4
Django==2.0.7
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-gis==0.13
drfapikey==0.0.3
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
Pygments==2.2.0
pylint==2.0.0
pytz==2018.5
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
